requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
greenlet==1.0.0
idna==2.10
msgpack==1.0.2
mysqlclient==2.0.3
PyJWT==2.0.1
pynvim==0.4.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
uuid==1.30
